% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_package_weather.report}
\alias{build_package_weather.report}
\title{Build weather report}
\usage{
build_package_weather.report(
  output = "pdf",
  args.vignette = list(),
  overwrite = FALSE,
  verbose = FALSE,
  devtools.ok = TRUE,
  require.data = TRUE
)
}
\arguments{
\item{output}{Character. Should report be pdf (default) or html.}

\item{args.vignette}{A named list defining arguments to be passed to the vignette.}

\item{overwrite}{A boolean indicating whether to overwrite existing compiled vignettes (`TRUE`) or not (`FALSE`, the default). If `FALSE`, the user will be prompted in case the vignette already exists.}

\item{verbose}{Logical. Should the function generate messages (default: TRUE)}

\item{devtools.ok}{A boolean indicating whether to build the vignette when the package is loading using devtools (`TRUE`, the default) or not (`FALSE`).}

\item{require.data}{A boolean indicating whether to build the vignette data must have been loaded (`TRUE`, the default) or not (`FALSE`).}
}
\description{
Build weather report
}
