% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_function_arg.location.weather}
\alias{check_function_arg.location.weather}
\title{Check weather station locations are valid.}
\usage{
check_function_arg.location.weather(location)
}
\value{
A character vector of weather station locations
}
\description{
Check weather station locations are valid.
}
\examples{
#Will work
check_function_arg.location.weather("jackal_hill")

#Won't work
#check_function_arg.location.weather("not_a_location")
}
