% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_function_arg.variable.weather}
\alias{check_function_arg.variable.weather}
\title{Check input variable exist in the weather dataset}
\usage{
check_function_arg.variable.weather(variable, .fill = TRUE)
}
\arguments{
\item{variable}{A character vector. Name of columns to select from the dataframe.}

\item{.fill}{Whether to automatically fill missing argument(s).}
}
\value{
A character vector of column names
}
\description{
Check input variable exist in the weather dataset
}
\examples{
check_function_arg.variable.weather(c("air_temp", "precip"))
}
