% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_weatherstation_metadata.all}
\alias{check_weatherstation_metadata.all}
\title{Check for meta-data issues when combining weather data}
\usage{
check_weatherstation_metadata.all(metadata)
}
\arguments{
\item{metadata}{Data frame. Metadata of a file to be checked.}
}
\value{
Data frame of meta-data.
}
\description{
Checks include:
1. Check data is from the same device (serial number)
2. Check data is collected at same measurement interval
3. Check data has been collected at the same locaation (lat/long)
}
\details{
Check will pass if data locations are within 10m.
This allows for uncertainty in coordinate and the possibility
that weather stations will need to be moved slightly over time.
}
