% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_weatherstation_metadata.file}
\alias{check_weatherstation_metadata.file}
\title{Check for meta-data issues in a single weather station file}
\usage{
check_weatherstation_metadata.file(metadata, verbose = TRUE)
}
\arguments{
\item{metadata}{Data frame. Metadata of a file to be checked.}

\item{verbose}{Logical. Should the function generate messages (default: TRUE)}
}
\value{
Data frame of meta-data
}
\description{
Checks include:
1. Check that weather station is located within Ngorongoro Crater
2. Check that there are no errors recorded in the meta-data
}
