% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_crater_weather.summary.table}
\alias{create_crater_weather.summary.table}
\title{Create summary statistics of weather}
\usage{
create_crater_weather.summary.table(
  resolution = "day",
  from = NULL,
  to = NULL,
  at = NULL,
  variable = c("air_temp", "precip"),
  station = NULL,
  location = NULL
)
}
\arguments{
\item{resolution}{Character. Resolution at which weather data will be summarised.
Is 'day' by default, but can take any argument accepted for 'by' in [seq.Date()].}

\item{from}{The start date of a date range using the format "YYYY-MM-DD" or "YYYY/MM/DD". Combined with argument 'to' to create date range. Will default to earliest database observation date unless otherwise specified. Cannot be used with argument 'at'.}

\item{to}{The end date of a date range using the format "YYYY-MM-DD" or "YYYY/MM/DD". Combined with argument 'from' to create date range. Will default to latest database observation date unless otherwise specified. Cannot be used with argument 'at'.}

\item{at}{A single date using the format "YYYY-MM-DD" or "YYYY/MM/DD". Cannot be used with arguments 'from' and 'to'.}

\item{variable}{A character vector. Name of columns to select from the dataframe.}

\item{station}{A character vector. Vector of weather station names.}

\item{location}{Character vector. Location of weather station where data should be used. Can be 'ngoitokitok' or 'acacia'.}
}
\description{
Create summary statistics of weather
}
