% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_crater_weather.table}
\alias{create_crater_weather.table}
\title{Create a table of weather data in a given date range}
\usage{
create_crater_weather.table(
  from = NULL,
  to = NULL,
  at = NULL,
  variable = c("air_temp", "precip"),
  station = NULL,
  location = NULL
)
}
\value{
Tibble of weather data
}
\description{
Create a table of weather data in a given date range
}
\examples{
#Get temp data from jua station
create_crater_weather.table(system.file("extdata/working_weather", package = "NgoroWeather"),
                              variable = "temp", station = "jua")

}
