% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_station_activity.log}
\alias{create_station_activity.log}
\title{Create a table of weather station activity}
\usage{
create_station_activity.log(input.folder, ...)
}
\arguments{
\item{input.folder}{File path. Folder containing all input files.}
}
\value{
List containing two items: activity_log (summary of active
periods for each weather station) and cleaning_log (table of days
when station was cleaned)
}
\description{
Create a table of weather station activity
}
