% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_tbl_metadata}
\alias{create_tbl_metadata}
\title{Create a table of the meta-data attribute of a dataframe.}
\usage{
create_tbl_metadata(input.tbl)
}
\arguments{
\item{input.tbl}{Weather data table to extract meta-data.}
}
\value{
Tibble of meta-data information.
}
\description{
Create a table of the meta-data attribute of a dataframe.
}
\examples{
#Example weather data collected from IZW
test_data <- load_data_weatherstation.file(
  system.file("extdata/working_weather/weather_Mlima",
  "weather_data_test1.xlsx",
  package = "NgoroWeather"))

#Extract metadata
create_tbl_metadata(test_data)
}
