% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_weather_raw.table}
\alias{create_weather_raw.table}
\title{Create a table of all weather data at 30 min intervals}
\usage{
create_weather_raw.table(input.folder)
}
\value{
Tibble of meta-data information.
}
\description{
Create a table of all weather data at 30 min intervals
}
\examples{
#Example weather data collected from IZW
test_data <- create_weather_raw.table(system.file("extdata/working_weather",
                                          package = "NgoroWeather"))
}
