% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch_weather_fn}
\alias{fetch_weather_fn}
\title{Generate summary statistics column(s) for weather variable(s).}
\usage{
fetch_weather_fn(
  from = NULL,
  to = NULL,
  at = NULL,
  station = NULL,
  location = NULL,
  variable = NULL,
  fn = NULL,
  suffix = NULL
)
}
\arguments{
\item{from}{The start date of a date range using the format "YYYY-MM-DD" or "YYYY/MM/DD". Combined with argument 'to' to create date range. Will default to earliest database observation date unless otherwise specified. Cannot be used with argument 'at'.}

\item{to}{The end date of a date range using the format "YYYY-MM-DD" or "YYYY/MM/DD". Combined with argument 'from' to create date range. Will default to latest database observation date unless otherwise specified. Cannot be used with argument 'at'.}

\item{at}{A single date using the format "YYYY-MM-DD" or "YYYY/MM/DD". Cannot be used with arguments 'from' and 'to'.}

\item{station}{A character vector. Vector of weather station names.}

\item{location}{Character vector. Location of weather station where data should be used. Can be 'ngoitokitok' or 'acacia'.}

\item{variable}{A character vector. Name of columns to select from the dataframe.}

\item{fn}{A vector of functions or function names.}

\item{suffix}{Character string. Added to the end of column name to differentiate different summary statistics columns.}
}
\value{
A tibble with summary statistic column(s).
Summary statistics are calculated separately for each station/location.
}
\description{
NOTE: This function is best suited for applying custom functions, or apply a single function to multiple.
weather variables. For common statistics (mean, min, max, sd)
applied to temperature and rainfall see wrapper functions like [hyenaR::fetch_weather_temp.mean()].
}
\examples{
#Summary stat from station jua in October 2021
data.frame(from = "2021-10-01", to = "2021-10-31",
           station = "jua") |>
           dplyr::mutate(fetch_weather_fn(from = from, to = to, station = station,
                         variable = c("temp", "rain"), fn = mean, suffix = "mean"))

#Apply a custom function to find difference between temperature min/max
data.frame(from = "2021-10-01", to = "2021-10-31",
           station = "jua") |>
           dplyr::mutate(fetch_weather_fn(from = from, to = to, station = station,
                         variable = "temp", fn = ~{max(.) - min(.)}, suffix = "diff"))
}
