% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch_weather_rain.mean}
\alias{fetch_weather_rain.mean}
\title{Fetch mean rainfall from a weather station during a given period}
\usage{
fetch_weather_rain.mean(
  from = NULL,
  to = NULL,
  at = NULL,
  station = NULL,
  location = NULL
)
}
\value{
A tibble with mean rainfall Mean rainfall is calculated separately for each station/location.
}
\description{
NOTE: To apply custom functions please see [hyenaR::fetch_weather_fn()].
}
\examples{
#Summary stat from station jua in October 2021
data.frame(from = "2021-10-01", to = "2021-10-31",
           station = "jua") |>
           dplyr::mutate(fetch_weather_rain.mean(from = from, to = to, station = station))
}
