% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch_weather_temp.sd}
\alias{fetch_weather_temp.sd}
\title{Fetch standard deviation of temperature from a weather station during a given period}
\usage{
fetch_weather_temp.sd(
  from = NULL,
  to = NULL,
  at = NULL,
  station = NULL,
  location = NULL
)
}
\value{
A tibble with standard deviation of temperature.
Standard deviation is calculated separately for each station/location.
}
\description{
NOTE: To apply custom functions please see [hyenaR::fetch_weather_fn()].
}
\examples{
#Determine mean of temperature and rainfall from station jua in October 2021
data.frame(from = "2021-10-01", to = "2021-10-31",
           station = "jua") |>
           dplyr::mutate(fetch_weather_temp.sd(from = from, to = to, station = station))
}
