% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_data_excel}
\alias{load_data_excel}
\title{Load all sheets from an excel file (.xls or .xslx).}
\usage{
load_data_excel(excel.path, ...)
}
\arguments{
\item{excel.path}{File path to .xlsx or .xls file.}

\item{...}{Arguments passed to readxl::read_excel.}
}
\value{
A list of tibbles. Each list item represents a sheet in the excel
spreadsheet
}
\description{
Load all sheets from an excel file (.xls or .xslx).
}
\examples{
#Example data collected from IZW is stored with the package for example
load_data_excel(system.file("extdata/working_weather/weather_Mlima",
                                          "weather_data_test1.xlsx",
                                          package = "NgoroWeather"))

#Can pass arguments like ignoring columns names using ...
load_data_excel(system.file("extdata/working_weather/weather_Mlima",
                                          "weather_data_test1.xlsx",
                                          package = "NgoroWeather"), col_names = FALSE)
}
