% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_data_weatherstation.all}
\alias{load_data_weatherstation.all}
\title{Load and compile all files from a single weather station into R}
\usage{
load_data_weatherstation.all(input.folder, verbose = TRUE)
}
\arguments{
\item{input.folder}{File path. Folder containing all input files.}

\item{verbose}{Logical. Should the function generate messages (default: TRUE)}
}
\value{
A list with two items: weather data & station meta data
}
\description{
'tz' argument is used when converting excel date times.
}
\examples{
#Example data collected from IZW is stored with the package for example
load_data_weatherstation.all(system.file("extdata/working_weather/weather_Mlima",
package = "NgoroWeather"))
}
