% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_data_weatherstation.file}
\alias{load_data_weatherstation.file}
\title{Load single weather station file into R.}
\usage{
load_data_weatherstation.file(
  excel.path,
  tz = "Africa/Dar_es_Salaam",
  verbose = TRUE
)
}
\arguments{
\item{excel.path}{File path to .xlsx or .xls file.}

\item{tz}{Time zone used of date-time object. See ?timezones for more details on time zones in R.}

\item{verbose}{Logical. Should the function generate messages (default: TRUE)}
}
\value{
A list with two items: weather data & station meta data
}
\description{
'tz' argument is used when converting excel date times (default = "Africa/Dar_es_Salaam").
}
\examples{
#Example data collected from IZW is stored with the package for example
load_data_weatherstation.file(system.file("extdata/weather_Mlima",
                                          "weather_data_test1.xlsx",
                                          package = "NgoroWeather"))
}
