% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multifetch.R
\name{multifetch_weather_rain.summary}
\alias{multifetch_weather_rain.summary}
\title{Fetch basic summary statistics of rainfall for given weather station(s) on given date.}
\usage{
multifetch_weather_rain.summary(
  from = NULL,
  to = NULL,
  at = NULL,
  station = NULL,
  location = NULL
)
}
\value{
A tibble with mean, max, min, and standard deviation of rainfall in the focal period.
}
\description{
Fetch basic summary statistics of rainfall for given weather station(s) on given date.
}
\examples{
data.frame(from = "2021-10-01", to = "2021-10-31",
           station = "jua") |>
           dplyr::mutate(multifetch_weather_rain.summary(from = from, to = to, station = station))
}
