% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_function_arg.variable.weather}
\alias{check_function_arg.variable.weather}
\title{Check input variable exist in the weather dataset}
\usage{
check_function_arg.variable.weather(variable, .fill = TRUE)
}
\value{
A character vector of column names
}
\description{
Check input variable exist in the weather dataset
}
\examples{
check_function_arg.variable.weather(c("temp", "rain"))
}
