% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch_weather_fn}
\alias{fetch_weather_fn}
\title{Generate summary statistics column(s) for weather variable(s).}
\usage{
fetch_weather_fn(
  from = NULL,
  to = NULL,
  at = NULL,
  station = NULL,
  location = NULL,
  variable = NULL,
  fn = NULL,
  suffix = NULL
)
}
\value{
A tibble with summary statistic column(s).
Summary statistics are calculated separately for each station/location.
}
\description{
NOTE: This function is best suited for applying custom functions, or apply a single function to multiple.
weather variables. For common statistics (mean, min, max, sd)
applied to temperature and rainfall see wrapper functions like [hyenaR::fetch_weather_temp.mean()].
}
\examples{
#Summary stat from station jua in October 2021
data.frame(from = "2021-10-01", to = "2021-10-31",
           station = "jua") |>
           dplyr::mutate(fetch_weather_fn(from = from, to = to, station = station,
                         variable = c("temp", "rain"), fn = mean, suffix = "mean"))

#Apply a custom function to find difference between temperature min/max
data.frame(from = "2021-10-01", to = "2021-10-31",
           station = "jua") |>
           dplyr::mutate(fetch_weather_fn(from = from, to = to, station = station,
                         variable = "temp", fn = ~{max(.) - min(.)}, suffix = "diff"))
}
