% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch_weather_temp.sd}
\alias{fetch_weather_temp.sd}
\title{Fetch standard deviation of temperature from a weather station during a given period}
\usage{
fetch_weather_temp.sd(
  from = NULL,
  to = NULL,
  at = NULL,
  station = NULL,
  location = NULL
)
}
\arguments{
\item{from}{The start date of a date range using the format "YYYY-MM-DD" or "YYYY/MM/DD". Combined with argument 'to' to create date range. Will default to earliest database observation date unless otherwise specified. Cannot be used with argument 'at'.}

\item{to}{The end date of a date range using the format "YYYY-MM-DD" or "YYYY/MM/DD". Combined with argument 'from' to create date range. Will default to latest database observation date unless otherwise specified. Cannot be used with argument 'at'.}

\item{at}{A single date using the format "YYYY-MM-DD" or "YYYY/MM/DD". Cannot be used with arguments 'from' and 'to'.}

\item{station}{A character vector. Vector of weather station names.}

\item{location}{Character vector. Location of weather station where data should be used. Can be 'ngoitokitok' or 'acacia'.}
}
\value{
A tibble with standard deviation of temperature.
Standard deviation is calculated separately for each station/location.
}
\description{
NOTE: To apply custom functions please see [hyenaR::fetch_weather_fn()].
}
\examples{
#Load data
load_package_database.weather(system.file("extdata/working_weather", package = "NgoroWeather"),
overwrite.db = "yes")

#Determine mean of temperature and rainfall from station jua in October 2021
data.frame(from = "2021-10-01", to = "2021-10-31",
           station = "jua") |>
           dplyr::mutate(fetch_weather_temp.sd(from = from, to = to, station = station))
}
