% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multifetch.R
\name{multifetch_weather_rain.summary}
\alias{multifetch_weather_rain.summary}
\title{Fetch basic summary statistics of rainfall for given weather station(s) on given date.}
\usage{
multifetch_weather_rain.summary(
  from = NULL,
  to = NULL,
  at = NULL,
  station = NULL,
  location = NULL
)
}
\arguments{
\item{from}{The start date of a date range using the format "YYYY-MM-DD" or "YYYY/MM/DD". Combined with argument 'to' to create date range. Will default to earliest database observation date unless otherwise specified. Cannot be used with argument 'at'.}

\item{to}{The end date of a date range using the format "YYYY-MM-DD" or "YYYY/MM/DD". Combined with argument 'from' to create date range. Will default to latest database observation date unless otherwise specified. Cannot be used with argument 'at'.}

\item{at}{A single date using the format "YYYY-MM-DD" or "YYYY/MM/DD". Cannot be used with arguments 'from' and 'to'.}

\item{station}{A character vector. Vector of weather station names.}

\item{location}{Character vector. Location of weather station where data should be used. Can be 'ngoitokitok' or 'acacia'.}
}
\value{
A tibble with mean, max, min, and standard deviation of rainfall in the focal period.
}
\description{
Fetch basic summary statistics of rainfall for given weather station(s) on given date.
}
\examples{
load_package_database.weather(system.file("extdata/working_weather",
package = "NgoroWeather"), overwrite.db = "yes")

data.frame(from = "2021-10-01", to = "2021-10-31",
           station = "jua") |>
           dplyr::mutate(multifetch_weather_rain.summary(from = from, to = to, station = station))
}
